% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateEnvelope.R
\name{climateEnvelope}
\alias{climateEnvelope}
\title{Create a climate envelope plot of one or multiple species}
\usage{
climateEnvelope(
  data,
  climate = "worldclim",
  model = "MI",
  res = 10,
  lon = "x",
  lat = "y",
  rcp = 60,
  year = 50,
  extent = NA,
  method = "MCP",
  path = ""
)
}
\arguments{
\item{data}{RasterLayer, SpatialPointsDataFrame or data.frame}

\item{climate}{character. One of Worldclim, Chelsa or ISIMIP2b}

\item{model}{character. Should be one of "AC", "BC", "CC", "CE", "CN", "GF", "GD", "GS", "HD", 
"HG", "HE", "IN", "IP", "MI", "MR", "MC", "MP", "MG", or "NO".}

\item{res}{Valid resolutions are 0.5, 2.5, 5, and 10 (minutes of a degree). In the case of res=0.5, you must also provide a lon and lat argument for a tile; for the lower resolutions global data will be downloaded.}

\item{lon}{Longitude argument for a tile, only applicable if res = 0.5}

\item{lat}{Latitude argument for a tile, only applicable if res = 0.5}

\item{rcp}{character. RCP scenario for future climate data, should be one of 26, 45, 60 or 85.}

\item{year}{numeric. Year of future climate scenario, should be one of 50, 70.}

\item{extent}{specify extent of area considered}

\item{method}{character. One of CH (convex hull) or MCP (minimum convex polygon) or MFB (personal method).}

\item{path}{Character. Path name indicating where to store the data. 
Default is the current working directory.}
}
\value{
ggplot2 object
}
\description{
Read a SpatialPointsDataFrame or dataframe with species location data and 
return a plot with the current and future climate envelope
}
\examples{
\dontrun{
library(climateNiche)
data(Passer_domesticus)

climateEnvelope(data=Passer_domesticus, method="MCP")

climateEnvelope(data=Passer_domesticus, method="CH")

climateEnvelope(data=Passer_domesticus, method="MFB")
}
}
