% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateNiche-package.R
\docType{package}
\name{climateNiche package}
\alias{climateNiche package}
\alias{climateNichepackage}
\title{R Package for analysing the climatic niche of a species}
\description{
R Package for analysing the climatic niche of a species
}
\details{
R package for analysing the climatic niche of a species by utilising information from IUCN, GBIF, eBird and Worldclim
}
\author{
RS-eco
}
\keyword{package}
