% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateNiche-package.R
\docType{data}
\name{Equus_quagga}
\alias{Equus_quagga}
\title{Point locations of Equus_quagga}
\format{
\code{sp::SpatialPointsDataFrame}
}
\description{
SpatialPointsDataFrame with GBIF occurrences of the Plain's zebra Equus quagga
}
\details{
This SpatialPointsDataFrame contains point occurrences of the plain'S zebra Equus quagga
derived from GBIF using the rgbif package.
}
