% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateNiche-package.R
\docType{data}
\name{Passer_castanopterus}
\alias{Passer_castanopterus}
\title{Species range of Passer castanopterus}
\format{
\code{raster::RasterLayer}
}
\description{
RasterLayer with IUCN expert range of the Somali sparrow Passer castanopterus
}
\details{
This RasterLayer depicts the expert range of the Somali sparrow Passer castanopterus
derived from IUCN expert species range polygons.
}
