% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateNiche-package.R
\docType{data}
\name{Passer_domesticus}
\alias{Passer_domesticus}
\title{Species range of Passer domesticus}
\format{
\code{raster::RasterLayer}
}
\description{
RasterLayer with IUCN expert range of the house sparrow Passer domesticus
}
\details{
This RasterLayer depicts the expert range of the house sparrow Passer domesticus 
derived from IUCN expert species range polygons.
}
