% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateNiche-package.R
\docType{data}
\name{Passer_motitensis}
\alias{Passer_motitensis}
\title{Species range of Passer motitensis}
\format{
\code{raster::RasterLayer}
}
\description{
RasterLayer with IUCN expert range of the Greater sparrow Passer motitensis
}
\details{
This RasterLayer depicts the expert range of the Greater sparrow Passer domesticus 
derived from IUCN expert species range polygons.
}
