% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateNiche-package.R
\docType{data}
\name{Passer_zarudnyi}
\alias{Passer_zarudnyi}
\title{Species range of Passer zarudnyi}
\format{
\code{raster::RasterLayer}
}
\description{
RasterLayer with IUCN expert range of the Asian desert sparrow Passer zarudnyi
}
\details{
This RasterLayer depicts the expert range of the Asian desert sparrow Passer zarudnyi 
derived from IUCN expert species range polygons.
}
