% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSp.R
\name{plotSp}
\alias{plotSp}
\title{Create plot of  Distribution or Range}
\usage{
plotSp(
  data,
  name = "",
  extent = NA,
  country = TRUE,
  cut = FALSE,
  grid = FALSE,
  save = FALSE,
  filename,
  dpi = 300,
  width = NA,
  height = NA,
  units = c("in", "cm", "mm")
)
}
\arguments{
\item{data}{RasterLayer, SpatialPointsDataFrame or data.frame}

\item{name}{character. Adds a name to the legend if provided.}

\item{extent}{Specify extent of your Map}

\item{country}{Should country outline be plotted or not}

\item{cut}{logical. Cut outline by extent}

\item{grid}{logical. Should grid lines and axes be drawn or not}

\item{save}{logical. If TRUE plot object is saved, see also the subsequent parameters.}

\item{filename}{character. File name to create on disk.}

\item{dpi}{numeric. Plot resolution.}

\item{width, height, units}{Plot size in units ("in", "cm", or "mm"). If not supplied, uses the size of current graphics device.}
}
\value{
ggplot2 object
}
\description{
Read a SpatialPointsDataFrame or normal dataframe and plot the coordinates
}
\examples{
\dontrun{
library(climateNiche)
data(Passer_domesticus)

plotSp(data=Passer_domesticus)
}
}
