% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptSDM.R
\name{sptSDM}
\alias{sptSDM}
\title{Spatio-temporal species distribution model}
\usage{
sptSDM(species_data, env_data, tres = "months")
}
\arguments{
\item{species_data}{A SpatialPointsDataFrame.}

\item{env_data}{A raster stack of environmental data.}

\item{tres}{A time variable.}
}
\value{
spatio-temporal model of \code{species_data}
}
\description{
Function to create a spatio-temporal species distribution model
}
\examples{
sptSDM(species_data, env_data, tres="months")
}
