% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrestrialGBIF.R
\name{terrestrialGBIF}
\alias{terrestrialGBIF}
\title{Get GBIF data for a terrestrial species}
\usage{
terrestrialGBIF(
  species = "Equus quagga",
  genus = NULL,
  limit = 50000,
  hasCoordinate = TRUE,
  path = getwd(),
  overwrite = FALSE
)
}
\arguments{
\item{species}{Latin name of species.}

\item{genus}{Genus of species}

\item{limit}{Maximum number of records to obtain from GBIF. Note that there is a hard maximum of 200,000, which is calculated as the limit+start, so start=199,000 and limit=2000 won't work}

\item{hasCoordinate}{logical. Return only occurence records with lat/long data (TRUE) or all records (FALSE, default).}

\item{path}{File path, where to save the species data}

\item{overwrite}{logical. If TRUE, fill will be overwritten if it exists.}
}
\value{
SpatialPointsDataFrame with the occurrence of the species
}
\description{
Function to extract GBIF data and save it to file.
Output further does not contain any outliers and only files on land.
}
\examples{
\dontrun{
terrestrialGBIF(species="Equus quagga", limit=100)
}
}
